airflow/requirements.txt
# requirements.txt - Python dependencies for Airflow 3.1 DAGs

# Core dependencies - Airflow 3.1 with Python 3.12
apache-airflow[celery,postgres,redis,kubernetes]==3.1.0

# Providers (Airflow 3.1 compatible versions)
apache-airflow-providers-cncf-kubernetes>=8.5.0
apache-airflow-providers-postgres>=5.12.0
apache-airflow-providers-amazon>=8.30.0
apache-airflow-providers-google>=10.22.0
apache-airflow-providers-http>=4.13.0
apache-airflow-providers-slack>=8.8.0
apache-airflow-providers-ssh>=3.13.0
apache-airflow-providers-common-sql>=1.18.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
sqlalchemy>=2.0.0

# Cloud SDKs
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.18.0

# API clients
requests>=2.31.0
httpx>=0.24.0

# Data validation
pydantic>=2.0.0
great-expectations>=0.18.0

# Monitoring and observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
prometheus-client>=0.17.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
pendulum>=3.0.0

# Testing (for local development)
pytest>=7.4.0
pytest-cov>=4.1.0